Reapply test plan operation changes
